+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+ - Added new tool `mfd_multi_brute` - MIFARE DESfire / UL-C key recovery (@iceman1001)
  - Fixed `hf emrtd info` segfault on some platforms (@doegox)
  - Fixed `hf emrtd info` when offline (@doegox)
  - Fixed `commands.json` generation (@doegox)